fix(phase1): consolidate duplicated PBKDF2+AES-256-GCM into secureStorage.js
